refactor: 컴포넌트를 폴더로 분리
