+++ src/components/Input/Input.scss
@@ -1,4 +1,4 @@
-@import "../index.scss";
+@import "../../index.scss";
 
 $white: #fff;
 $gray: #9b9b9b;
