+++ src/components/Main/Home.scss
